Remove the C++ fxr_resolver wrappers
fxr_resolver.cpp held thin C++ wrappers (marshalling pal::string_t to and
from the pal_char_t* C API) over the fxr_resolver_* functions in
fxr_resolver.c. Remove them and have the remaining C++ callers use the C
API directly:

- hostfxr_resolver.cpp uses fxr_search_location and fxr_resolver_try_get_path
  directly (with direct utils.h/error_codes.h includes that were previously
  pulled in transitively through fxr_resolver.h).
- The load_fxr_and_get_delegate template (used by ijwhost and comhost) moves
  to its own header, load_fxr_and_get_delegate.h, and calls the C resolver
  functions. Its fxr handle is typed as pal_dll_t.

fxr_resolver.h is now a pure C header and fxr_resolver is C-only (no
fxr_resolver.cpp), which also drops it from the static singlefilehost
sources.
