+++ src/installer/managed/Microsoft.NET.HostModel/AppHost/HostWriter.cs
@@ -38,7 +38,7 @@ namespace Microsoft.NET.HostModel.AppHost
 
         public class DotNetSearchOptions
         {
-            // Keep in sync with fxr_resolver::search_location in fxr_resolver.h
+            // Keep in sync with fxr_search_location in fxr_resolver.h
             [Flags]
             public enum SearchLocation : byte
             {
