+++ src/native/corehost/apphost/static/CMakeLists.txt
@@ -30,7 +30,6 @@ set(SOURCES
     ./hostfxr_resolver.cpp
     ./hostpolicy_resolver.cpp
     ../../hostpolicy/static/coreclr_resolver.cpp
-    ../../fxr_resolver.cpp
     ../../fxr_resolver.c
     ../../corehost.cpp
 )
